refactor cmds into subcmd package names and files
