+++ pkg/cli/cmd/topaz/console.go
@@ -1,4 +1,4 @@
-package cmd
+package topaz
 
 import (
 	"strings"
