+++ pkg/cli/cmd/topaz/stop.go
@@ -1,4 +1,4 @@
-package cmd
+package topaz
 
 import (
 	"fmt"
@@ -7,6 +7,7 @@ import (
 
 	"github.com/aserto-dev/topaz/pkg/cc/config"
 	"github.com/aserto-dev/topaz/pkg/cli/cc"
+	"github.com/aserto-dev/topaz/pkg/cli/cmd/common"
 	"github.com/aserto-dev/topaz/pkg/cli/dockerx"
 	"github.com/fatih/color"
 )
@@ -71,7 +72,7 @@ func (cmd *StopCmd) Run(c *cc.CommonCtx) error {
 	// empty running config
 	c.Config.Running = cc.RunningConfig{}
 
-	if err := c.SaveContextConfig(CLIConfigurationFile); err != nil {
+	if err := c.SaveContextConfig(common.CLIConfigurationFile); err != nil {
 		fmt.Fprintln(os.Stderr, err.Error())
 		os.Exit(1)
 	}
